DXE-259 start end fields should be defined explicitly as int64
Merge in DEVEXP/terraform-provider-akamai from feature/DXE-259-start-end-fields-should-be-defined-explicitly-as-int64 to feature/sp-cloudlets2

+++ pkg/providers/cloudlets/data_akamai_cloudlets_api_prioritization_match_rule.go
@@ -240,8 +240,8 @@ func getMatchRulesAP(matchRules []interface{}) (cloudlets.MatchRules, error) {
 		matchRule := cloudlets.MatchRuleAP{
 			Name:               getStringValue(matchRuleMap, "name"),
 			Type:               cloudlets.MatchRuleTypeAP,
-			Start:              getIntValue(matchRuleMap, "start"),
-			End:                getIntValue(matchRuleMap, "end"),
+			Start:              getInt64Value(matchRuleMap, "start"),
+			End:                getInt64Value(matchRuleMap, "end"),
 			Matches:            matches,
 			MatchURL:           getStringValue(matchRuleMap, "match_url"),
 			PassThroughPercent: getFloat64PtrValue(matchRuleMap, "pass_through_percent"),
