+++ pkg/providers/cloudlets/data_akamai_cloudlets_visitor_prioritization_match_rule.go
@@ -240,8 +240,8 @@ func getMatchRulesVP(matchRules []interface{}) (cloudlets.MatchRules, error) {
 		matchRule := cloudlets.MatchRuleVP{
 			Name:               getStringValue(matchRuleMap, "name"),
 			Type:               cloudlets.MatchRuleTypeVP,
-			Start:              getIntValue(matchRuleMap, "start"),
-			End:                getIntValue(matchRuleMap, "end"),
+			Start:              getInt64Value(matchRuleMap, "start"),
+			End:                getInt64Value(matchRuleMap, "end"),
 			Matches:            matches,
 			MatchURL:           getStringValue(matchRuleMap, "match_url"),
 			PassThroughPercent: getFloat64PtrValue(matchRuleMap, "pass_through_percent"),
